requirements.txt
Django==2.2.13
djangorestframework==3.12.2
pygame==2.0.0
requests==2.25.0
